move v2.event to v2.core.event
