+++ stripe/_object_classes.py
@@ -153,7 +153,7 @@ V2_OBJECT_CLASSES = {
     stripe.v2.billing.MeterEvent.OBJECT_NAME: stripe.v2.billing.MeterEvent,
     stripe.v2.billing.MeterEventAdjustment.OBJECT_NAME: stripe.v2.billing.MeterEventAdjustment,
     stripe.v2.billing.MeterEventSession.OBJECT_NAME: stripe.v2.billing.MeterEventSession,
-    stripe.v2.Event.OBJECT_NAME: stripe.v2.Event,
-    stripe.v2.EventDestination.OBJECT_NAME: stripe.v2.EventDestination,
+    stripe.v2.core.Event.OBJECT_NAME: stripe.v2.core.Event,
+    stripe.v2.core.EventDestination.OBJECT_NAME: stripe.v2.core.EventDestination,
     # V2 Object classes: The end of the section generated from our OpenAPI spec
 }
