+++ stripe/_stripe_client.py
@@ -27,7 +27,7 @@ from stripe._stripe_response import StripeResponse
 from stripe._util import _convert_to_stripe_object, get_api_mode, deprecated  # noqa: F401
 from stripe._webhook import Webhook, WebhookSignature
 from stripe._event import Event
-from stripe.v2._event import EventNotification
+from stripe.v2.core._event import EventNotification
 
 from typing import Any, Dict, Optional, Union, cast
 from typing_extensions import TYPE_CHECKING
