+++ stripe/events/_v1_billing_meter_error_report_triggered_event.py
@@ -7,7 +7,7 @@ from stripe._stripe_object import StripeObject
 from stripe._stripe_response import StripeResponse
 from stripe._util import get_api_mode
 from stripe.billing._meter import Meter
-from stripe.v2._event import Event, EventNotification, RelatedObject
+from stripe.v2.core._event import Event, EventNotification, RelatedObject
 from typing import Any, Dict, List, Optional, cast
 from typing_extensions import Literal, override
