+++ stripe/events/_v1_billing_meter_no_meter_found_event.py
@@ -4,7 +4,7 @@ from stripe._api_mode import ApiMode
 from stripe._api_requestor import _APIRequestor
 from stripe._stripe_object import StripeObject
 from stripe._stripe_response import StripeResponse
-from stripe.v2._event import Event, EventNotification
+from stripe.v2.core._event import Event, EventNotification
 from typing import Any, Dict, List, Optional, cast
 from typing_extensions import Literal, override
