+++ stripe/events/_v2_core_event_destination_ping_event.py
@@ -3,8 +3,8 @@
 from stripe._stripe_client import StripeClient
 from stripe._stripe_object import StripeObject
 from stripe._util import get_api_mode
-from stripe.v2._event import Event, EventNotification, RelatedObject
-from stripe.v2._event_destination import EventDestination
+from stripe.v2.core._event import Event, EventNotification, RelatedObject
+from stripe.v2.core._event_destination import EventDestination
 from typing import Any, Dict, cast
 from typing_extensions import Literal, override
