+++ stripe/v2/__init__.py
@@ -1,12 +1,5 @@
 from stripe.v2._list_object import ListObject as ListObject
 from stripe.v2._amount import Amount as Amount, AmountParam as AmountParam
-from stripe.v2._event import (
-    EventNotification as EventNotification,
-    UnknownEventNotification as UnknownEventNotification,
-    RelatedObject as RelatedObject,
-    Reason as Reason,
-    ReasonRequest as ReasonRequest,
-)
 
 
 # The beginning of the section generated from our OpenAPI spec
@@ -14,6 +7,4 @@ from stripe.v2 import billing as billing, core as core
 from stripe.v2._billing_service import BillingService as BillingService
 from stripe.v2._core_service import CoreService as CoreService
 from stripe.v2._deleted_object import DeletedObject as DeletedObject
-from stripe.v2._event import Event as Event
-from stripe.v2._event_destination import EventDestination as EventDestination
 # The end of the section generated from our OpenAPI spec
