+++ stripe/v2/core/__init__.py
@@ -1,6 +1,19 @@
 # -*- coding: utf-8 -*-
-# File generated from our OpenAPI spec
+
+from stripe.v2.core._event import (
+    EventNotification as EventNotification,
+    RelatedObject as RelatedObject,
+    Reason as Reason,
+    ReasonRequest as ReasonRequest,
+)
+
+# The beginning of the section generated from our OpenAPI spec
+from stripe.v2.core._event import Event as Event
+from stripe.v2.core._event_destination import (
+    EventDestination as EventDestination,
+)
 from stripe.v2.core._event_destination_service import (
     EventDestinationService as EventDestinationService,
 )
 from stripe.v2.core._event_service import EventService as EventService
+# The end of the section generated from our OpenAPI spec
